Changed structure of folders
