+++ src/components/organisms/Rectangle/Rectangle.style.js
