+++ src/components/organisms/Trapeze/Trapeze.style.js
