+++ src/components/organisms/Triangle/Triangle.js
@@ -9,15 +9,15 @@ import Basic from 'assets/images/GeometricShapes/triangleBasic.png';
 import { TriangleContainer } from './Triangle.style';
 import { GeometricButton } from 'components/atoms/GeometricButton/GeometricButton.style';
 import { SubmitContainer } from 'components/atoms/GeometricButton/GeometricButton.style';
-import ResultTable from '../ResultTable/ResultTable';
+import ResultTable from 'components/molecules/ResultTable/ResultTable';
 import { ReactComponent as SvgBasic } from 'assets/images/GeometricShapes/triangleBasic.svg';
 import { ReactComponent as SvgIsoceles } from 'assets/images/GeometricShapes/triangleIsosceles.svg';
 import { ReactComponent as SvgEquilateral } from 'assets/images/GeometricShapes/triangleEquilateral.svg';
 import { ReactComponent as SvgRectangular } from 'assets/images/GeometricShapes/triangleRectangular.svg';
 import DynamicInputForImage from 'components/atoms/DynamicInputForImage/DynamicInputForImage';
 import { wrongDataAlert } from 'helpers/Swal';
-import InputWithUnits from '../InputWithUnits/InputWithUnits';
-import TriangleFunctions from 'helpers/TriangleFunctions';
+import InputWithUnits from 'components/molecules/InputWithUnits/InputWithUnits';
+import TriangleFunctions from 'helpers/GeometricFunctions/TriangleFunctions';
 import { countUndefined, transformToDecIfNeeded } from 'helpers/Helpers';
 import { radToDeg } from 'helpers/Helpers';
 import { degToRad } from 'helpers/Helpers';
