+++ src/components/organisms/Triangle/Triangle.style.js
