+++ src/helpers/GeometricFunctions/TrapezeFunctions.js
